fix: move downloads fixtures to fixtures folder
